Migrates Toolbar to own file
