+++ src/App.ts
@@ -2,7 +2,7 @@ import { MeiosisCell, MeiosisViewComponent } from 'meiosis-setup/types'
 import { Chess } from 'chess.js'
 import { State } from './State'
 import { Game } from './Game.ts'
-import { Toolbar } from './meiosis.ts'
+import { Toolbar } from './Toolbar.ts'
 import { Login } from './Login.ts'
 import { HistoryService, PageService } from './Services'
 import { UserProfile } from './User.ts'
