+++ src/meiosis.ts
@@ -11,39 +11,12 @@ import '@fortawesome/fontawesome-free/css/all.css'
 import alertify, { notify } from 'alertifyjs'
 import 'alertifyjs/build/css/alertify.css'
 import 'alertifyjs/build/css/themes/semantic.css'
-import { UserIcon } from './User.ts'
 
 alertify.defaults.transition = 'zoom'
 // alertify.defaults.theme.ok = 'ui positive button'
 // alertify.defaults.theme.cancel = 'ui black button'
 
-export const Toolbar = ({ state, update }) =>
-  m('#toolbar.component', {}, [
-    m('', {}, state.page),
-    m(
-      'button',
-      {
-        onclick: () => update({ page: 'Home' }),
-      },
-      'Home'
-    ),
-    m(
-      'button',
-      {
-        onclick: () => update({ page: 'Game' }),
-      },
-      'Game'
-    ),
-    state.user
-      ? UserIcon({ state, update })
-      : m(
-          'button',
-          {
-            onclick: () => update({ page: 'Login' }),
-          },
-          'Login'
-        ),
-  ])
+
 
 export const History = cell =>
   m(
