+++ src/user.css
@@ -1,15 +1,4 @@
-#toolbar {
-    position: relative;
-    height: 60px;
-    background-color: #606060;
-    padding: 10px; /* Adjust padding as needed */
-  }
-  
 .profile-icon {
-    position: absolute;
-
-    top: 5px;
-    right: 5px;
     width: 50px;
     height: 50px;
     border-radius: 50%;
